LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/static-router-subresource.https-expected.txt


PASS Subresource load not matched with URLPattern condition
FAIL Subresource load matched with URLPattern condition assert_equals: expected "Network\n" but got "arfcy"
FAIL Subresource cross origin load matched with URLPattern condition via constructed object assert_equals: expected 0 but got 2
FAIL Subresource load matched with ignoreCase URLPattern condition assert_equals: expected "Network\n" but got "ncqdg"
PASS Subresource load matched without ignoreCase URLPattern condition
FAIL Subresource load matched with URLPattern condition via URLPatternCompatible assert_equals: expected "Network\n" but got "mkjnr"
FAIL Subresource cross origin load not matched with URLPattern condition via URLPatternCompatible assert_equals: expected 1 but got 2
FAIL Subresource load matched with URLPattern condition via string assert_equals: expected "Network\n" but got "rfdaj"
FAIL Subresource cross origin load not matched with URLPattern condition via string assert_equals: expected 1 but got 2
PASS Subresource load matched with RequestMode condition
FAIL Subresource load matched with the nested `or` condition assert_equals: expected "Network\n" but got "ytihv"
FAIL Subresource load matched with the next `or` condition assert_equals: expected "Network\n" but got "lzkmm"
PASS Subresource load not matched with `or` condition
FAIL Subresource load matched with the cache source rule assert_equals: expected "From cache" but got ""
FAIL Subresource load did not match with the cache and fallback to the network assert_equals: expected "Network\n" but got "wytpc"
FAIL Subresource load matched with the cache source, with specifying the cache name assert_equals: expected "From cache" but got ""
FAIL Subresource load should not match with the not condition assert_equals: expected "xirxd" but got "{\"error\": {\"code\": 404, \"message\": \"404\"}}"
PASS Subresource load should match with a file other than not
